+++ CHANGELOG.md
@@ -16,11 +16,13 @@ Highlights are marked with a pancake 🥞
 ### Changed
 
 - Refactor mock `Node` implementation to use `StorageProvider` traits [#383](https://github.com/p2panda/p2panda/pull/383) `rs`
+- Deserialize from string and u64 for `LogId` and `SeqNum` [#401](https://github.com/p2panda/p2panda/pull/401) `rs`
 
 ### Fixed
 
 - Set log id default to `0` [#398](https://github.com/p2panda/p2panda/pull/398) `rs`
 - Fix iterator implementations for `SeqNum` and `LogId` [#404](https://github.com/p2panda/p2panda/pull/404) `rs`
+- Fix system schema CDDL definitions #[393](https://github.com/p2panda/p2panda/pull/393) `rs`
 
 ## [0.4.0]